+++ py7zr/py7zr.py
@@ -448,7 +448,14 @@ class SevenZipFile:
                 self.worker.register_filelike(f.id, outfilename)
                 target_files.append((outfilename, f.file_properties()))
         for target_dir in sorted(target_dirs):
-            os.mkdir(target_dir)
+            if not os.path.exists(target_dir):
+                os.mkdir(target_dir)
+            elif os.path.isdir(target_dir):
+                pass
+            elif os.path.isfile(target_dir):
+                raise("Directory name is existed as a normal file.")
+            else:
+                raise
         self.worker.extract(self.fp)
         for b, t in target_sym:
             b.seek(0)